Rename to AvoidShadowingTypeParameters
