+++ example/all.yaml
@@ -31,6 +31,7 @@ linter:
     - avoid_returning_null_for_void
     - avoid_returning_this
     - avoid_setters_without_getters
+    - avoid_shadowing_type_parameters
     - avoid_single_cascade_in_expression_statements
     - avoid_slow_async_io
     - avoid_types_as_parameter_names
@@ -104,7 +105,6 @@ linter:
     - prefer_void_to_null
     - public_member_api_docs
     - recursive_getters
-    - shadow_type_parameters
     - slash_for_doc_comments
     - sort_constructors_first
     - sort_pub_dependencies
