+++ lib/src/rules.dart
@@ -32,6 +32,7 @@ import 'package:linter/src/rules/avoid_returning_null.dart';
 import 'package:linter/src/rules/avoid_returning_null_for_void.dart';
 import 'package:linter/src/rules/avoid_returning_this.dart';
 import 'package:linter/src/rules/avoid_setters_without_getters.dart';
+import 'package:linter/src/rules/avoid_shadowing_type_parameters.dart';
 import 'package:linter/src/rules/avoid_single_cascade_in_expression_statements.dart';
 import 'package:linter/src/rules/avoid_slow_async_io.dart';
 import 'package:linter/src/rules/avoid_types_as_parameter_names.dart';
@@ -106,7 +107,6 @@ import 'package:linter/src/rules/pub/package_names.dart';
 import 'package:linter/src/rules/pub/sort_pub_dependencies.dart';
 import 'package:linter/src/rules/public_member_api_docs.dart';
 import 'package:linter/src/rules/recursive_getters.dart';
-import 'package:linter/src/rules/shadow_type_parameters.dart';
 import 'package:linter/src/rules/slash_for_doc_comments.dart';
 import 'package:linter/src/rules/sort_constructors_first.dart';
 import 'package:linter/src/rules/sort_unnamed_constructors_first.dart';
@@ -167,6 +167,7 @@ void registerLintRules() {
     ..register(new AvoidReturnTypesOnSetters())
     ..register(new AvoidReturningThis())
     ..register(new AvoidSettersWithoutGetters())
+    ..register(new AvoidShadowingTypeParameters())
     ..register(new AvoidSingleCascadeInExpressionStatements())
     ..register(new AvoidSlowAsyncIo())
     ..register(new AvoidTypesAsParameterNames())
@@ -239,7 +240,6 @@ void registerLintRules() {
     ..register(new PublicMemberApiDocs())
     ..register(new PubPackageNames())
     ..register(new RecursiveGetters())
-    ..register(new ShadowTypeParameters())
     ..registerDefault(new SlashForDocComments())
     ..register(new SortConstructorsFirst())
     ..register(new SortPubDependencies())
